Story/ragkit 17 introduce embedding set (#2)
* [#RAGKIT-17] Introduce Embedding Set

We currently track the state of ingestion by checking if a library has
chunks at all, and, if yes, all of them have an embedding. This was ok
as quick solution, but it makes it a bit harder to display which
libraries are in the ingestion pipeline, and if they failed, and, - if
yes -where.

Let’s introduce EmbeddingSet, which hold all the chunks of a library,
and the state of the library. We connect it with the library through a
join model LibraryEmbeddingSet

It also paves the way for future polymorphism and might allow to keep
the chunks of the current ingestion until a re-ingestion has passed.

https://bitcrowd.atlassian.net/browse/RAGKIT-17

* created migrations

* models in place

* factories

* just before changing the workers

* GenerateEmbeddingsWorker & Test updated

* linting...

* q and none

* Update lib/exmeralda/topics/ingest_library_worker.ex

Co-authored-by: Joel Koch <joel@joelkoch.dev>

* final touches

* make credo happy again

* intoduced current_embedding_set

---------

Co-authored-by: Joel Koch <joel@joelkoch.dev>